requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
groq>=0.4.0
pdfplumber>=0.10.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
pytesseract>=0.3.10
Pillow>=10.0.0
striprtf>=0.0.26
odfpy>=1.4.1
email-validator>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
